LayoutTests/accessibility/mac/area-with-aria-label.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/accessibility-helper.js"></script>
<script src="../../resources/js-test.js"></script>
</head>
<body>

<map name="forwardbackward">
  <area id="btn_prev" role="button" alt="" tabindex="0" aria-label="Previous" shape="rect" coords="2,35,21,47" href="#">
  <area id="btn_next" role="button" alt="" tabindex="0" aria-label="Next" shape="rect" coords="56,35,75,47" href="#">
</map>
<img id="fb" src="#" border="1" width="76" height="76" usemap="#forwardbackward" alt="">

<script>
let output = "This tests that aria-label works on area elements.\n\n";

if (window.accessibilityController) {
    output += expect("accessibilityController.accessibleElementById('btn_prev').description", "'AXDescription: Previous'");
    output += expect("accessibilityController.accessibleElementById('btn_next').description", "'AXDescription: Next'");

    debug(output);
}
</script>
</body>
</html>
