LayoutTests/accessibility/mac/area-with-aria-label-expected.txt


This tests that aria-label works on area elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.description is 'AXDescription: Previous'
PASS accessibilityController.focusedElement.description is 'AXDescription: Next'
PASS successfullyParsed is true

TEST COMPLETE
